Refactor authentication stuff into own module.
